이미지 파이썬으로 수정하는 기능 활성화
Former-commit-id: 0f48fe03734fcbed2046406fe6f8f4298bd2e13b

+++ front-react/src/App.js
@@ -99,7 +99,7 @@ export default class App extends React.Component {
         auth: true,
         user_id: splited[0],
         authority: splited[1],
-        tempLimit: parseInt(splited[3]),
+        tempLimit: parseFloat(splited[3]),
         tempType: parseFloat(splited[2])
       });
     }
@@ -123,7 +123,7 @@ export default class App extends React.Component {
           auth: true,
           user_id: splited[0],
           authority: splited[1],
-          tempLimit: parseInt(splited[3]),
+          tempLimit: parseFloat(splited[3]),
           tempType: parseFloat(splited[2])
         });
       }
